requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pypdf==5.9.0
python-docx==1.2.0
